chore: ruff lint + format cleanup + hardware conftest fix
Clears 184 pre-existing ruff errors accumulated since v0.5.0. CI has been
red on main for 24 consecutive runs, all blocked at the ruff-check step
before any tests could run.

Changes:
- ruff check --fix --unsafe-fixes (71 mechanical fixes: import sort, unused
  imports, UP037 quoted-annotations now unnecessary under
  `from __future__ import annotations`, zip→itertools.pairwise).
- ruff format (81 files: multi-line reformat for long DH-param tables,
  nested fixtures, argument lists).
- Hand-fixed remaining 6: `try/except/pass` → `contextlib.suppress`,
  en-dash→hyphen in comments, one overlong fixture string split, one
  docstring trimmed.
- Fix tests/hardware/conftest.py: `pytest_collection_modifyitems` was
  adding `skip` to *every* collected item, not just ones marked
  `@pytest.mark.hardware`. Latent bug from 9f6f74b — masked because CI
  never got past lint. The root tests/conftest.py already filters by
  the hardware marker correctly, so the hardware-dir hook is redundant;
  replaced with a marker-only stub.

No behavior changes: src/ edits are stylistic only. Tests still pass
individually; the only runtime issue surfaced during verification is that
unit tests that hit `load_context` against a fixture manifest cause
`dai.Device()` to block probing local OAK-D hardware — a pre-existing
test-isolation issue, out of scope for this cleanup. CI runners have no
OAK-D and will fail fast there.

+++ cli/src/robot_md/gripper_silhouette.py
@@ -40,12 +40,14 @@ def expected_points(
     # Since we don't have full tool-frame orientation here, approximate by
     # offsetting ±5mm in y and ±3mm in z from tip. Good enough for depth
     # matching at ~500mm standoff where pixel footprint >> 5mm.
-    offsets = np.array([
-        [0, 0, 0],
-        [0, 5, 0],
-        [0, -5, 0],
-        [0, 0, 3],
-    ])
+    offsets = np.array(
+        [
+            [0, 0, 0],
+            [0, 5, 0],
+            [0, -5, 0],
+            [0, 0, 3],
+        ]
+    )
     return tip_base[None, :] + offsets
 
 
@@ -70,11 +72,11 @@ def find_in_depth(
     cx, cy = float(K[0, 2]), float(K[1, 2])
     if center[2] <= 0:
         return None, 0.0
-    u0 = int(round(fx * center[0] / center[2] + cx))
-    v0 = int(round(fy * center[1] / center[2] + cy))
+    u0 = round(fx * center[0] / center[2] + cx)
+    v0 = round(fy * center[1] / center[2] + cy)
 
     # Convert search radius from mm to pixels at that depth.
-    radius_px = int(round((search_radius_mm / center[2]) * fx))
+    radius_px = round((search_radius_mm / center[2]) * fx)
     radius_px = max(radius_px, 8)
 
     h, w = depth_frame.shape[:2]
